checkout and order

+++ HalceraAPI.Model/Price.cs
@@ -17,14 +17,14 @@ namespace HalceraAPI.Models
         [Key]
         public int Id { get; set; }
         [Required]
-        public double? Amount { get; set; }
+        public decimal? Amount { get; set; }
         [Required]
         public Currency? Currency { get; set; }
 
         /// <summary>
         /// indicate if price is discounted
         /// </summary>
-        public double? DiscountAmount { get; set; }
+        public decimal? DiscountAmount { get; set; }
 
         public int? ProductId { get; set; }
         [ForeignKey(nameof(ProductId))]
